# Speed up importing mqtt platforms